+++ tests/testthat/test-garnish_plot.R
@@ -1,6 +1,7 @@
 testthat::test_that("garnish_plot", {
 
-  on.exit(list.files(pattern = "(antigen.garnish.*\\.pdf)|(antigen.garnish_example_jaffa)") %>% file.remove)
+  if (identical(Sys.getenv("TESTTHAT"), "true"))
+    on.exit(list.files(pattern = "(antigen.garnish.*\\.pdf)|(antigen.garnish_example_jaffa)") %>% file.remove)
 
   # load test data
     dt <- data.table::fread("http://get.rech.io/antigen.garnish_example_jaffa_output.txt")
@@ -12,4 +13,3 @@ testthat::test_that("garnish_plot", {
     testthat::expect_equal(list.files(pattern = "antigen.garnish.*score.*\\.pdf") %>% length, 2)
 
 })
-