+++ tests/testthat/test-garnish_summary.R
@@ -11,15 +11,15 @@ testthat::test_that("garnish_summary", {
      testthat::expect_equal(dto$variants, 2)
      testthat::expect_equal(ncol(dto), 21)
 
-     dt <- data.table::fread("http://get.rech.io/antigen.garnish_example_output.txt") %>%
-      garnish_fitness %>%
-      .[, fitness_score := 5]
+     dt <- data.table::fread("http://get.rech.io/antigen.garnish_pureCN_example_output.txt")
 
     dto2 <- garnish_summary(dt)
 
-    testthat::expect_equal(dto2$nmers, 154)
-    testthat::expect_equal(dto2$mhc_binders_class_I, 3)
-    testthat::expect_equal(dto2$variants, 2)
-    testthat::expect_equal(ncol(dto2), 23)
+    testthat::expect_equal(dto2$nmers, 269)
+    testthat::expect_equal(dto2$mhc_binders_class_I, 8)
+    testthat::expect_equal(dto2$variants, 4)
+    testthat::expect_equal(ncol(dto2), 24)
+    testthat::expect_equal(dto2$garnish_score %>% signif(digits = 3), 74900)
+    testthat::expect_equal(dto2$fitness_scores_class_I %>% signif(digits = 3), 16.2)
 
     })